chore: drop 222: Remove KotlinCompat
