+++ kotlin/src/com/google/idea/blaze/kotlin/sync/BlazeKotlinProjectConfigurator.java
@@ -17,15 +17,78 @@ package com.google.idea.blaze.kotlin.sync;
 
 import com.google.idea.sdkcompat.kotlin.KotlinCompat.KotlinProjectConfiguratorBase;
 import com.intellij.openapi.module.Module;
+import com.intellij.openapi.project.Project;
+import com.intellij.openapi.roots.DependencyScope;
+import com.intellij.openapi.roots.ExternalLibraryDescriptor;
+import com.intellij.psi.PsiElement;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+import org.jetbrains.kotlin.config.ApiVersion;
+import org.jetbrains.kotlin.config.LanguageFeature;
+import org.jetbrains.kotlin.idea.base.projectStructure.ModuleSourceRootGroup;
 import org.jetbrains.kotlin.idea.configuration.ConfigureKotlinStatus;
+import org.jetbrains.kotlin.idea.configuration.KotlinProjectConfigurator;
+import org.jetbrains.kotlin.idea.projectConfiguration.LibraryJarDescriptor;
+import org.jetbrains.kotlin.platform.TargetPlatform;
+import org.jetbrains.kotlin.platform.jvm.JvmPlatforms;
+
+import java.util.Collection;
 
 /** A kotlin project configurator to suppress "Configure Kotlin" banner. */
-public class BlazeKotlinProjectConfigurator extends KotlinProjectConfiguratorBase {
+public class BlazeKotlinProjectConfigurator extends KotlinProjectConfiguratorBase implements KotlinProjectConfigurator {
 
   @NotNull
-  @Override
   public ConfigureKotlinStatus getStatus(@NotNull Module module) {
     return ConfigureKotlinStatus.CONFIGURED;
   }
+
+  @NotNull
+  @Override
+  public String getName() {
+    return "Bazel";
+  }
+
+  @NotNull
+  @Override
+  public String getPresentableText() {
+    return "Bazel";
+  }
+
+  @NotNull
+  @Override
+  public TargetPlatform getTargetPlatform() {
+    return JvmPlatforms.INSTANCE.getUnspecifiedJvmPlatform();
+  }
+
+  @Override
+  public void addLibraryDependency(
+      @NotNull Module module,
+      @NotNull PsiElement psiElement,
+      @NotNull ExternalLibraryDescriptor externalLibraryDescriptor,
+      @NotNull LibraryJarDescriptor libraryJarDescriptor,
+      @NotNull DependencyScope dependencyScope) {}
+
+  @Override
+  public void changeGeneralFeatureConfiguration(
+      @NotNull Module module,
+      @NotNull LanguageFeature languageFeature,
+      @NotNull LanguageFeature.State state,
+      boolean b) {}
+
+  @Override
+  public void configure(@NotNull Project project, @NotNull Collection<Module> collection) {}
+
+  @NotNull
+  @Override
+  public ConfigureKotlinStatus getStatus(@NotNull ModuleSourceRootGroup moduleSourceRootGroup) {
+    return getStatus(moduleSourceRootGroup.getBaseModule());
+  }
+
+  @Override
+  public void updateLanguageVersion(
+      @NotNull Module module,
+      @Nullable String s,
+      @Nullable String s1,
+      @NotNull ApiVersion apiVersion,
+      boolean b) {}
 }
